pyproject.toml
[tool.poetry]
name = "inter-fully-connected"
version = "0.1.0"
description = "New neural network with every nodes interconnected"
authors = ["Loïc <b.loic38@laposte.net>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "inter_fully_connected"}]

[tool.poetry.scripts]
ifc = 'inter_fully_connected.cli:cli'

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.0.0"
torchvision = "^0.15.1"
pytorch-lightning = "^2.0.0"
ipykernel = "^6.22.0"
scipy = "^1.10.1"
tensorboard = "^2.12.0"
click = "^8.1.3"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
